Fix: [AEA-4168] - 200 response for duplicate item (#384)
## Summary
2 Conflicting responses are returned for the same item if a 409 error is
found due to a duplicate Task.ID. The expected 409 response is returned,
however a 200 is returned with for the same Task.ID noting that the data
is not committed. However, we want to maintain the response of other
items and only replace the ones with a duplicate Task ID.

https://nhsd-jira.digital.nhs.uk/browse/AEA-4168

- Routine Change

### Details

Add any summary information of what is in the change. **Remove this line
if you have nothing to add.**