Update: [AEA-4180] - Update custom PSU transformation lambda to handle DispenseComplete (#544)
## Summary

- Routine Change

### Details

Two additions.

#### Update the schema for prescription items

The item entry in an incoming cPSU request can contain a
`completedStatus` field, which can take the values `["Cancelled",
"Expired", "NotDispensed", "Collected"]`. The schema has been updated to
include this option (it is not mandatory, and may not be given).

#### Handle excess messages

External providers will send additional messages after we consider a
prescription to be completed. We may receive updates to the cPSU
translation layer that mark an item as both `status:
"DispensingComplete"`, and `completedStatus: "<Cancelled, Expired,
NotDispensed>"`. In these cases, we will already have received a
previous update flagging the item as failed in some way, so we want to
ignore such messages. To handle that, we simply do not populate the
entry array with the relevant item.

Note that downstream interfaces need to behave well with empty `entry`
arrays, for the case where all items are ignored.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: anthony-nhs <121869075+anthony-nhs@users.noreply.github.com>
Co-authored-by: Tom Smith <69810386+TomSmithCoding@users.noreply.github.com>
Co-authored-by: Jack Spagnoli <jack.spagnoli1@nhs.net>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Kris Szlapa <kris.szlapa1@nhs.net>