New: [AEA-5199] - Setup notifications flow (#1546)
## Summary

- :sparkles: New Feature

### Details

This PR is to add the various components needed for the NHS notify
integration to the PSU.

There are four new things to deploy:
- EventBridge scheduler
- Lambda function to handle the processing, triggered by the scheduler
- SQS (eventually to be populated by the existing PSU function)
- Notifications state Dynamo database

Initially, the only requirement is that the lambda function is executed
every minute. The PSU will not push anything to SQS yet. The lambda will
only log that it has been run.