New: [AEA-5202] - Filter for sites with notifications allowed or denied (#1574)
## Summary

- :sparkles: New Feature

### Details

Adds a function on the producer side of EPS notifications that filters
out sites that are blocked, whilst only allowing through enabled
suppliers and sites. Case insensitive.

We want to maintain these lists in version control, so I've simply added
them as static data to the helper function file. There are also three
test cases for us to use in testing:
- Enabled test ODS code: `FA565`
- Enabled test supplier name: `Internal Test System`
- Blocked ODS code: `A83008`