Fix: [AEA-0000] - Add an upper time limit for the lambda runtime (#1846)
## Summary

- Routine Change

### Details

Lambdas have a timeout - heavy traffic was causing the lambdas to
continually process the queue, and exceed that timeout as they never
observed an empty one. This just adds a time limit to how long a lambda
can be processing the queue before it gives up and yields. Note that the
old behaviour works, but throws a timeout error that we don't want to
see.