Fix: [AEA-0000] - Use the APIM required header, rather than the one notify want us to use. Also, parameterise notify enable flags (#1884)
## Summary

- :warning: Potential issues that might be caused by this change

### Details

In order to test our callback endpoint, we need to be able to actually
use it. Since there's a mismatch between notify's spec, and APIM's
capabilities, I'm making it align with APIM for now just so we can see
how things behave.

Also, moves the notification enabling flags to deployment parameters,
and disables them both by default for prod/int