Chore: [AEA-5562] - Notification latency logging (#1930)
## Summary

- Routine Change

### Details

When the callback lambda processes an update, log the NHS number and the
original request ID for that item.

I also moved the type for the notification state table records to common
types, and use it in the callback handling.

After some discussion with Pete, we've decided to update the dynamo
behaviour so that rather than having one record per patient per ODS
code, which gets re-used by subsequent notification requests, we're
going to have a new record per patient, per PSU request ID. This does
require a little finagling, since it means updating the keys of the
table, but since we're not in prod yet we don't need to worry about any
migration headaches. We will need to delete the old PSU notification
state tables during deployments, though.