Fix: [AEA-0000] - Make target attribute "name" optional (#2118)
## Summary

- Routine Change

### Details

The Notify callbacks are not getting though. They're being denied
because of missing target attribute headers:

<img width="1095" height="1115" alt="image"
src="https://github.com/user-attachments/assets/0ec85f11-ba01-45e0-bb1d-0c168a49a641"
/>

I'm fairly confident that the issue is because we are requiring a name
target attribute header (this is necessary for the other endpoints in
the PSU), but the API key header we're using for the callback doesn't
seem to have a rule that injects this (the problem was hidden from me
because I accidentally had Bearer token auth still enabled in my
postman, which does inject this... oops). Chatted with Ant and we have a
solution, which is to make this an optional header
([here](https://github.com/NHSDigital/eps-prescription-status-update-api/blob/main/packages/specification/eps-prescription-status-update-api.yaml#L458))
and manually check it in the other lambdas