Fix: [AEA-0000] - Update how the callback is handled. (#2132)
## Summary

- Routine Change

### Details

The callback lambda currently kinda works, but not really. It rejects
requests that lack a request ID, and notify do not provide one, so that
needs to be fixed.

Also, I realised I never unit tested the new callback type, so I did
that.

Additionally, the logging of incoming updates is not really in line with
how statuses actually flow. I have tweaked it to be more generalised.