Fix: [AEA-5722] - Fix cooldown checking bug in notify processor lambda (#2152)
## Summary

- Routine Change

### Details

A small number of patients are getting lots of notifications. 

The cooldown check was only grabbing one record from the table - I'd
thought that if you sort by ODS code, then the latest record would be at
the top. This is not true, so we can be checking against old records
when newer ones exist.

In some rare cases (approx 1 in 10,000), a patient may get a LOT of PSU
requests about their medications. In one case study, a patient had 61
distinct PSU requests pertaining to 6 prescription IDs, of which 24
requests were ready to collect. All of that was from one ODS code. This
resulted in 5 SQS messages being posted ~5 minutes apart. However, when
the cooldown check came for subsequent requests, it always passed. I
believe this is because the cooldown check is essentially a `return
true` check once there are enough records.

This PR fixes this bug by instead grabbing all records, and then taking
the maximum timestamp and checking if that is within the cooldown
period.