Fix: [AEA-5743] - Fix the bypassing of the retry logic in the auth token exchange (#2209)
## Summary

- Routine Change

### Details

I had this line in the auth token axios request:
```
validateStatus: () => true
```
This meant that axios would resolve the promise for all status codes,
which I was using to catch failed logins and report them. However, this
meant that the retry logic for the interceptor was never being
triggered, and so failed logins were only ever attempted once.